+++ packages/css/src/transformCss.ts
@@ -1,5 +1,3 @@
-import './ahocorasick.d';
-
 import { getVarName } from '@vanilla-extract/private';
 import cssesc from 'cssesc';
 import AhoCorasick from 'ahocorasick';
@@ -149,6 +147,11 @@ class Stylesheet {
       return;
     }
     if (root.type === 'keyframes') {
+      root.rule = Object.fromEntries(
+        Object.entries(root.rule).map(([keyframe, rule]) => {
+          return [keyframe, this.transformProperties(rule)];
+        }),
+      );
       this.keyframesRules.push(root);
 
       return;
@@ -193,10 +196,8 @@ class Stylesheet {
     conditions: Array<string>,
     { isLayer }: { isLayer?: boolean } = {},
   ) {
-    // Run `pixelifyProperties` before `transformVars` as we don't want to pixelify CSS Vars
-    const rule = this.transformVars(
-      this.transformContent(this.pixelifyProperties(cssRule.rule)),
-    );
+    // Run `transformProperties` before `transformVars` as we don't want to pixelify CSS Vars
+    const rule = this.transformVars(this.transformProperties(cssRule.rule));
     const selector = this.transformSelector(cssRule.selector);
 
     if (!this.currConditionalRuleset) {
@@ -218,16 +219,18 @@ class Stylesheet {
   }
 
   addRule(cssRule: CSSRule) {
-    // Run `pixelifyProperties` before `transformVars` as we don't want to pixelify CSS Vars
-    const rule = this.transformVars(
-      this.transformContent(this.pixelifyProperties(cssRule.rule)),
-    );
+    // Run `transformProperties` before `transformVars` as we don't want to pixelify CSS Vars
+    const rule = this.transformVars(this.transformProperties(cssRule.rule));
     const selector = this.transformSelector(cssRule.selector);
 
     this.rules.push({
       selector,
       rule,
     });
+  }
+
+  transformProperties(cssRule: CSSPropertiesWithVars) {
+    return this.transformContent(this.pixelifyProperties(cssRule));
   }
 
   pixelifyProperties(cssRule: CSSPropertiesWithVars) {