+++ packages/integration/src/addFileScope.test.ts
@@ -1,34 +1,35 @@
 import { outdent } from 'outdent';
 import { addFileScope } from './addFileScope';
 
-test('should add missing fileScope', () => {
-  const source = outdent`
-  import {style} from '@vanilla-extract/css';
-
-  export const myStyle = style({});
-  `;
-
-  expect(
-    addFileScope({
-      source,
-      rootPath: '/the-root',
-      filePath: '/the-root/app/app.css.ts',
-      packageName: 'my-package',
-    }),
-  ).toMatchInlineSnapshot(`
-    "
-        import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
-        setFileScope("app/app.css.ts", "my-package");
-        import {style} from '@vanilla-extract/css';
+describe('ESM', () => {
+  test('should add missing fileScope', () => {
+    const source = outdent`
+    import {style} from '@vanilla-extract/css';
 
     export const myStyle = style({});
-        endFileScope();
-      "
-  `);
-});
+    `;
 
-test('should update existing fileScope', () => {
-  const source = outdent`
+    expect(
+      addFileScope({
+        source,
+        rootPath: '/the-root',
+        filePath: '/the-root/app/app.css.ts',
+        packageName: 'my-package',
+      }),
+    ).toMatchInlineSnapshot(`
+      "
+            import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
+            setFileScope("app/app.css.ts", "my-package");
+            import {style} from '@vanilla-extract/css';
+
+      export const myStyle = style({});
+            endFileScope();
+          "
+    `);
+  });
+
+  test('should update existing fileScope', () => {
+    const source = outdent`
     import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
     setFileScope("some-weird-file", "some-weird-package");
     import {style} from '@vanilla-extract/css';
@@ -37,76 +38,168 @@ test('should update existing fileScope', () => {
     endFileScope();
     `;
 
-  expect(
-    addFileScope({
-      source,
-      rootPath: '/the-root',
-      filePath: '/the-root/app/app.css.ts',
-      packageName: 'my-package',
-    }),
-  ).toMatchInlineSnapshot(`
-    "import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
-    setFileScope("app/app.css.ts", "my-package");
-    import {style} from '@vanilla-extract/css';
+    expect(
+      addFileScope({
+        source,
+        rootPath: '/the-root',
+        filePath: '/the-root/app/app.css.ts',
+        packageName: 'my-package',
+      }),
+    ).toMatchInlineSnapshot(`
+      "import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
+      setFileScope("app/app.css.ts", "my-package");
+      import {style} from '@vanilla-extract/css';
 
-    export const myStyle = style({});
-    endFileScope();"
-  `);
-});
+      export const myStyle = style({});
+      endFileScope();"
+    `);
+  });
 
-test('should update existing fileScope with newlines', () => {
-  const source = outdent`
+  test('should update existing fileScope with newlines', () => {
+    const source = outdent`
       import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
       setFileScope(
-        "some-weird-file", 
+        "some-weird-file",
         "some-weird-package"
       );
       import {style} from '@vanilla-extract/css';
-  
+
       export const myStyle = style({});
       endFileScope();
       `;
 
-  expect(
-    addFileScope({
-      source,
-      rootPath: '/the-root',
-      filePath: '/the-root/app/app.css.ts',
-      packageName: 'my-package',
-    }),
-  ).toMatchInlineSnapshot(`
-    "import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
-    setFileScope("app/app.css.ts", "my-package");
-    import {style} from '@vanilla-extract/css';
+    expect(
+      addFileScope({
+        source,
+        rootPath: '/the-root',
+        filePath: '/the-root/app/app.css.ts',
+        packageName: 'my-package',
+      }),
+    ).toMatchInlineSnapshot(`
+      "import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
+      setFileScope("app/app.css.ts", "my-package");
+      import {style} from '@vanilla-extract/css';
 
-    export const myStyle = style({});
-    endFileScope();"
-  `);
-});
+      export const myStyle = style({});
+      endFileScope();"
+    `);
+  });
 
-test('should handle namespaced filescope calls', () => {
-  const source = outdent`
+  test('should handle namespaced filescope calls', () => {
+    const source = outdent`
     import * as vanillaFileScope from "@vanilla-extract/css/fileScope";
     vanillaFileScope.setFileScope("some-weird-file");
     import {style} from '@vanilla-extract/css';
 
     export const myStyle = style({});
     vanillaFileScope.endFileScope();
-  `;
+    `;
 
-  expect(
-    addFileScope({
-      source,
-      rootPath: '/the-root',
-      filePath: '/the-root/app/app.css.ts',
-      packageName: 'my-package',
-    }),
-  ).toMatchInlineSnapshot(`
-    "import * as vanillaFileScope from "@vanilla-extract/css/fileScope";
-    vanillaFileScope.setFileScope("app/app.css.ts", "my-package");
-    import {style} from '@vanilla-extract/css';
+    expect(
+      addFileScope({
+        source,
+        rootPath: '/the-root',
+        filePath: '/the-root/app/app.css.ts',
+        packageName: 'my-package',
+      }),
+    ).toMatchInlineSnapshot(`
+          "import * as vanillaFileScope from "@vanilla-extract/css/fileScope";
+          vanillaFileScope.setFileScope("app/app.css.ts", "my-package");
+          import {style} from '@vanilla-extract/css';
 
-    export const myStyle = style({});
-    vanillaFileScope.endFileScope();"
-  `);
+          export const myStyle = style({});
+          vanillaFileScope.endFileScope();"
+      `);
+  });
+});
+
+describe('CJS', () => {
+  test('should add missing fileScope', () => {
+    const source = outdent`
+    const _css require('@vanilla-extract/css');
+
+    var myStyle = _css.style({});
+    exports.myStyle = myStyle;
+    `;
+
+    expect(
+      addFileScope({
+        source,
+        rootPath: '/the-root',
+        filePath: '/the-root/app/app.css.ts',
+        packageName: 'my-package',
+      }),
+    ).toMatchInlineSnapshot(`
+      "
+          const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
+          __vanilla_filescope__.setFileScope("app/app.css.ts", "my-package");
+          const _css require('@vanilla-extract/css');
+
+      var myStyle = _css.style({});
+      exports.myStyle = myStyle;
+          __vanilla_filescope__.endFileScope();
+        "
+    `);
+  });
+
+  test('should update existing fileScope', () => {
+    const source = outdent`
+    const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
+    __vanilla_filescope__.setFileScope("some-weird-file", "some-weird-package");
+    const _css = require('@vanilla-extract/css');
+
+    var myStyle = _css.style({});
+    exports.myStyle = myStyle;
+    __vanilla_filescope__.endFileScope();
+    `;
+
+    expect(
+      addFileScope({
+        source,
+        rootPath: '/the-root',
+        filePath: '/the-root/app/app.css.ts',
+        packageName: 'my-package',
+      }),
+    ).toMatchInlineSnapshot(`
+      "const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
+      __vanilla_filescope__.setFileScope("app/app.css.ts", "my-package");
+      const _css = require('@vanilla-extract/css');
+
+      var myStyle = _css.style({});
+      exports.myStyle = myStyle;
+      __vanilla_filescope__.endFileScope();"
+    `);
+  });
+
+  test('should update existing fileScope with newlines', () => {
+    const source = outdent`
+      const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
+      __vanilla_filescope__.setFileScope(
+        "some-weird-file",
+        "some-weird-package"
+      );
+      const _css = require('@vanilla-extract/css');
+
+      const myStyle = _css.style({});
+      exports.myStyle = myStyle;
+      __vanilla_filescope__.endFileScope();
+      `;
+
+    expect(
+      addFileScope({
+        source,
+        rootPath: '/the-root',
+        filePath: '/the-root/app/app.css.ts',
+        packageName: 'my-package',
+      }),
+    ).toMatchInlineSnapshot(`
+      "const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
+      __vanilla_filescope__.setFileScope("app/app.css.ts", "my-package");
+      const _css = require('@vanilla-extract/css');
+
+      const myStyle = _css.style({});
+      exports.myStyle = myStyle;
+      __vanilla_filescope__.endFileScope();"
+    `);
+  });
 });