+++ site/webpack.config.js
@@ -69,7 +69,7 @@ module.exports = [
           ],
         },
         {
-          test: /\.vanilla\.css$/i,
+          test: /\.css$/i,
           use: [MiniCssExtractPlugin.loader, 'css-loader'],
         },
         {
@@ -108,7 +108,7 @@ module.exports = [
     target: 'node',
     externals: [
       nodeExternals({
-        allowlist: [/^react-syntax-highlighter\/dist\/esm/],
+        allowlist: [/^react-syntax-highlighter\/dist\/esm/, /^@docsearch\/css/],
       }),
     ],
     output: {
@@ -146,8 +146,8 @@ module.exports = [
           ],
         },
         {
-          test: /\.vanilla\.css$/i,
-          use: ['css-loader'],
+          test: /\.css$/i,
+          use: ['null-loader'],
         },
         {
           test: /\.mdx?$/,